added appbuilder and fixed icons
